some refactoring and comments
